support gamma correction in test, train stage too, fixed warp inference error in test.py
